lattest

+++ Backend/Domain Layer/Migrations/20231220133717_miginit.cs
@@ -114,7 +114,7 @@ namespace Domain_Layer.Migrations
                     StartDate = table.Column<DateTime>(type: "datetime2", nullable: false),
                     EndDate = table.Column<DateTime>(type: "datetime2", nullable: false),
                     Country = table.Column<string>(type: "nvarchar(max)", nullable: false),
-                    Budget = table.Column<int>(type: "int", nullable: false),
+                    Budget = table.Column<double>(type: "float", nullable: false),
                     Status = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     tenantName = table.Column<string>(type: "nvarchar(max)", nullable: false)
                 },
